+++ src/main/java/frc/robot/RobotContainer.java
@@ -19,7 +19,6 @@ import edu.wpi.first.math.controller.ProfiledPIDController;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
-import frc.robot.commands.GetRobotPosition;
 import frc.robot.commands.SetPipeline;
 import frc.robot.commands.LimeLight.GetLocationOfAprilTag;
 import edu.wpi.first.math.trajectory.Trajectory;
@@ -40,12 +39,13 @@ import edu.wpi.first.wpilibj2.command.SwerveControllerCommand;
 import edu.wpi.first.wpilibj2.command.button.CommandPS4Controller;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
 import edu.wpi.first.wpilibj2.command.button.Trigger;
-import frc.StaticConstants.BlingConstants;
 import frc.robot.Constants.OperatorConstants;
-import frc.robot.commands.DefaultDriveCommand;
+import frc.robot.StaticConstants.BlingConstants;
 import frc.robot.commands.balanceTest;
-import frc.robot.commands.sturdyBaseCommand;
 import frc.robot.commands.BlingCommands.BlingSetPattern;
+import frc.robot.commands.DriveTrainCommands.DefaultDriveCommand;
+import frc.robot.commands.DriveTrainCommands.GetRobotPosition;
+import frc.robot.commands.DriveTrainCommands.sturdyBaseCommand;
 import frc.robot.commands.ExtenderCommands.ExtendExtender;
 import frc.robot.commands.ExtenderCommands.RetractExtender;
 import frc.robot.commands.ShoulderCommands.ShoulderLower;
