+++ src/main/java/frc/robot/subsystems/ExtenderSubsystem.java
@@ -11,11 +11,11 @@ import com.revrobotics.CANSparkMaxLowLevel.MotorType;
 
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.StaticConstants.HardwareMap;
-import frc.StaticConstants.MaxMotorAmpsConstants;
 import frc.robot.Constants;
 import frc.robot.PositionValues;
 import frc.robot.RobotContainer;
+import frc.robot.StaticConstants.HardwareMap;
+import frc.robot.StaticConstants.MaxMotorAmpsConstants;
 
 //  This Subsystem is for our Extender Arm and will be powered by one Spark Neo550;
