+++ src/main/java/frc/robot/subsystems/ShoulderSubsystem.java
@@ -19,10 +19,10 @@ import com.revrobotics.RelativeEncoder;
 
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.StaticConstants.HardwareMap;
-import frc.StaticConstants.MaxMotorAmpsConstants;
 import frc.robot.Constants;
 import frc.robot.PositionValues;
+import frc.robot.StaticConstants.HardwareMap;
+import frc.robot.StaticConstants.MaxMotorAmpsConstants;
 
 /*  This Subsystem is for our Winch that raises and lowers the arm to certain 
  *  set positions.
